+++ libs/src/core/remote/handler.c
@@ -816,8 +816,8 @@ void arts_remote_handle_send_already_local(void *pack) {
 }
 
 void arts_remote_get_from_db(arts_guid_t edt_guid, arts_guid_t db_guid,
-                             unsigned int slot, unsigned int offset,
-                             unsigned int len, uint32_t flags,
+                             unsigned int slot, uint64_t offset,
+                             uint64_t len, uint32_t flags,
                              unsigned int rank) {
   struct arts_remote_get_put_packet_s packet;
   packet.edt_guid = edt_guid;
